requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.2; python_version >= '3.7'
brotli==1.0.9
certifi==2022.12.7; python_full_version >= '3.6.0'
cffi==1.15.1
charset-normalizer==2.1.1; python_full_version >= '3.6.0'
cryptography==38.0.4; python_full_version >= '3.6.0'
cssselect2==0.7.0; python_version >= '3.7'
deprecated==1.2.13; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
django-anymail==9.0
django-cors-headers==3.13.0
django-oauth-toolkit==2.2.0
django==4.1.4
djangorestframework-simplejwt==5.2.2
djangorestframework==3.14.0
et-xmlfile==1.1.0; python_full_version >= '3.6.0'
fonttools[woff]==4.38.0; python_version >= '3.7'
gunicorn==20.1.0
html5lib==1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
idna==3.4; python_version >= '3.5'
jwcrypto==1.4.2; python_full_version >= '3.6.0'
oauthlib==3.2.2; python_full_version >= '3.6.0'
openpyxl==3.0.10
pillow==9.3.0
pycparser==2.21
pydyf==0.5.0; python_version >= '3.7'
pyjwt==2.6.0; python_version >= '3.7'
pyphen==0.13.2; python_version >= '3.7'
pytz==2022.7
qrcode==7.3.1
requests==2.28.1; python_version >= '3.7' and python_version < '4'
setuptools==65.6.3; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.3; python_version >= '3.5'
tinycss2==1.2.1; python_version >= '3.7'
urllib3==1.26.13; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
weasyprint==57.1
webencodings==0.5.1
wrapt==1.14.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
xlsxwriter==3.0.3
zopfli==0.2.2
